UP appbar in specific component
